third_party/blink/web_tests/external/wpt/workers/dedicated-worker-in-data-url-context.window-expected.txt

This is a testharness.js-based test.
[FAIL] Create a dedicated worker in a data url frame
  assert_equals: expected "PASS" but got "Worker construction unexpectedly synchronously failed"
[FAIL] Create a dedicated worker in a data url dedicated worker
  assert_equals: expected "PASS" but got "Worker construction unexpectedly synchronously failed"
Harness: the test ran to completion.
